+++ src/ska_low_mccs_spshw/daq_receiver/daq_component_manager.py
@@ -19,6 +19,10 @@ from time import sleep
 from typing import Any, Callable, Final, Optional
 
 import numpy as np
+
+# This is introducing gRPC as a dependency here where really we want to be
+# agnostic to the underlying implementation.
+from grpc._channel import _InactiveRpcError as InactiveRPCError  # type: ignore
 from ska_control_model import CommunicationStatus, PowerState, ResultCode, TaskStatus
 from ska_low_mccs_daq_interface import DaqClient
 from ska_ser_skuid.client import SkuidClient  # type: ignore
@@ -50,6 +54,7 @@ class DaqComponentManager(TaskExecutorComponentManager):
         component_state_callback: Callable[..., None],
         received_data_callback: Callable[[str, str, str], None],
         daq_initialisation_retry_frequency: int = 5,
+        dedicated_bandpass_daq: bool = False,
     ) -> None:
         """
         Initialise a new instance of DaqComponentManager.
@@ -73,6 +78,9 @@ class DaqComponentManager(TaskExecutorComponentManager):
             received from a tile
         :param daq_initialisation_retry_frequency: Frequency at which daq
             initialisation in retried.
+        :param dedicated_bandpass_daq: Flag indicating whether this DaqReceiver
+            is dedicated exclusively to monitoring bandpasses. If true then
+            this DaqReceiver will attempt to automatically monitor bandpasses.
         """
         self._power_state_lock = threading.RLock()
         self._started_event = threading.Event()
@@ -81,6 +89,7 @@ class DaqComponentManager(TaskExecutorComponentManager):
         self._consumers_to_start: str = "Daqmodes.INTEGRATED_CHANNEL_DATA"
         self._receiver_started: bool = False
         self._daq_id = str(daq_id).zfill(3)
+        self._dedicated_bandpass_daq = dedicated_bandpass_daq
         self._configuration = {}
         if receiver_interface:
             self._configuration["receiver_interface"] = receiver_interface
@@ -93,7 +102,11 @@ class DaqComponentManager(TaskExecutorComponentManager):
         self._daq_client = DaqClient(daq_address)
         self._skuid_url = skuid_url
         self._daq_initialisation_retry_frequency = daq_initialisation_retry_frequency
-        self._stop_establishing_communication = False
+        # True initially to prevent bandpass monitoring starting before adminMode is set
+        self._stop_establishing_communication = True
+        threading.Thread(
+            target=self.establish_communication, args=[json.dumps(self._configuration)]
+        ).start()
 
         super().__init__(
             logger,
@@ -102,9 +115,10 @@ class DaqComponentManager(TaskExecutorComponentManager):
             power=None,
             fault=None,
         )
-        self._task_executor = TaskExecutor(max_workers=2)
+        # We've bumped this to 3 workers to allow for the bandpass monitoring.
+        self._task_executor = TaskExecutor(max_workers=3)
 
-    def restart_daq_if_active(self) -> None:
+    def restart_daq_if_active(self: DaqComponentManager) -> None:
         """Restart daq if consumers are active."""
 
         def stop_daq_completion_callback(
@@ -116,7 +130,7 @@ class DaqComponentManager(TaskExecutorComponentManager):
         ) -> None:
             """Update stop_daq command status.
 
-            Executes start_daq command if stop_daq succeedes.
+            Executes start_daq command if stop_daq succeeds.
 
             :param input_data: input data for start_daq command
             :param status: the status of the asynchronous task
@@ -163,7 +177,7 @@ class DaqComponentManager(TaskExecutorComponentManager):
             )
 
     def start_daq_completion_callback(
-        self,
+        self: DaqComponentManager,
         status: TaskStatus | None = None,
         progress: int | None = None,
         result: JSONData = None,
@@ -185,12 +199,13 @@ class DaqComponentManager(TaskExecutorComponentManager):
             daq_status: dict[str, Any] = json.loads(self.daq_status())
             self.logger.info("DaqStatus check results - %s", daq_status)
             if daq_status["Bandpass Monitor"] is True:
-                bandpass_monitor_thread = threading.Thread(
-                    target=self.start_bandpass_monitor,
-                    name="submit_task_for_bandpass_monitoring",
-                    args=[json.dumps({"plot_directory": "/tmp"})],
-                )
-                bandpass_monitor_thread.start()
+                # bandpass_monitor_thread = threading.Thread(
+                #     target=self.start_bandpass_monitor,
+                #     name="submit_task_for_bandpass_monitoring",
+                #     args=[json.dumps({"plot_directory": "/tmp"})],
+                # )
+                # bandpass_monitor_thread.start()
+                self.start_bandpass_monitor(json.dumps({"plot_directory": "/tmp"}))
         else:
             self.logger.warning(
                 "The TaskStatus for StartDaq command is not COMPLETED. "
@@ -208,31 +223,181 @@ class DaqComponentManager(TaskExecutorComponentManager):
             )  # noqa: E501
 
         self._stop_establishing_communication = False
-        configuration = json.dumps(self._configuration)
-        threading.Thread(
-            target=self.establish_communication, args=[configuration]
-        ).start()
 
-    def establish_communication(self, configuration: str) -> None:
+    def _check_comms(self: DaqComponentManager) -> dict[str, Any] | None:
+        """
+        Make a call to the backend and catch comms errors so we can reconnect.
+
+        Status was chosen for convenience as the result will be used anyway
+        if we're connected.
+
+        :return: The current status of the DaqReceiver if no comms errors.
+        """
+        try:
+            return json.loads(self.daq_status())
+        # TODO: Refactor this to catch general communication errors.
+        # Probably want daq-interface to translate gRPC exceptions to more
+        # general ones that we can catch here.
+        except (InactiveRPCError, ConnectionError):
+            # This try/except block should be refactored into a decorator.
+            # This will make Daq more resilient to the Chaos Monkey.
+            self.logger.warning("Unable to communicate with DaqServer. Reinitialising.")
+            # If the server pod crashed with a consumer running we need to
+            # clear this manually as stop_daq won't be callable.
+            # Worst case scenario the user is prompted to call StopDaq manually.
+            self._started_event.clear()
+            return None
+
+    def establish_communication(self: DaqComponentManager, configuration: str) -> None:
         """Establish communication with the DaqReceiver components.
+
+        Runs an endless thread to continually try to establish comms while
+        the device is ONLINE.
 
         :param configuration: Configuration string for daq initialisation
         """
-        while not self._stop_establishing_communication:
-            try:
-                response = self._daq_client.initialise(configuration)
-                self.logger.info(response["message"])
-                self._update_communication_state(CommunicationStatus.ESTABLISHED)
-                self.restart_daq_if_active()
-                break
-            except Exception as e:  # pylint: disable=broad-except
-                self.logger.error(
-                    "Caught exception in start_communicating: %s. "
-                    + "Retrying in %s secs",
-                    e,
-                    self._daq_initialisation_retry_frequency,
-                )
+        while True:
+            while not self._stop_establishing_communication:
+                current_status = self._check_comms()
+                if current_status is None:
+                    # current_status == None if comms dropped.
+                    self._reestablish_communication(configuration)
+                    # Give comms a moment to establish before checking again
+                    sleep(2)
+                    continue
+                if self._dedicated_bandpass_daq:
+                    self._check_bandpass_monitor(current_status)
                 sleep(self._daq_initialisation_retry_frequency)
+            while self._stop_establishing_communication:
+                sleep(self._daq_initialisation_retry_frequency)
+
+    def _check_bandpass_monitor(
+        self: DaqComponentManager, status: dict[str, Any]
+    ) -> None:
+        """Check bandpass monitor status and get running if necessary.
+
+        :param status: The current status of the DaqReceiver.
+        """
+        if not all(
+            [
+                self._is_bandpass_monitor_running(status),
+                self._is_integrated_channel_consumer_running(status),
+            ]
+        ):
+            self.logger.warning("Problem detected in bandpass monitor, fixing...")
+            self._get_bandpass_running()
+
+    def _get_bandpass_running(self: DaqComponentManager) -> None:
+        """
+        Get the bandpass monitor running if it isn't already.
+
+        This method is called when the monitoring thread detects that either
+        the consumer has stopped or the bandpass monitor itself has stopped.
+        It starts the INTEGRATED DATA consumer and starts the bandpass monitor with
+        `auto_handle_daq=True`.
+        This device-side handles starting the correct consumer, the daq-server side
+        handles any reconfiguration.
+        """
+
+        def _wait_for_status(status: str, value: str) -> None:
+            """
+            Wait for Daq to achieve a certain status.
+
+            Intended as a helper to wait for a consumer or bandpass monitor to start.
+            Linear increase to retry time.
+
+            :param status: The Daq status category to check.
+            :param value: The expected value of the status.
+            """
+            daq_status = str(json.loads(self.daq_status())[status])
+            retry_count = 0
+            while value not in daq_status:
+                retry_count += 1
+                sleep(retry_count)
+                daq_status = str(json.loads(self.daq_status())[status])
+                if retry_count > 5:
+                    self.logger.error(
+                        f"Failed to find {value} in DaqStatus[{status}]: {daq_status=}."
+                    )
+                    return
+            self.logger.debug(f"Found {value} in DaqStatus[{status}]: {daq_status=}.")
+
+        if not self._is_integrated_channel_consumer_running():
+            # start consumer
+            self.logger.info(
+                "Auto starting INTEGRATED DATA consumer for bandpass monitoring."
+            )
+            self.start_daq(modes_to_start="INTEGRATED_CHANNEL_DATA")
+            _wait_for_status(
+                status="Running Consumers", value=str(["INTEGRATED_CHANNEL_DATA", 5])
+            )
+
+        if not self._is_bandpass_monitor_running():
+            bandpass_args = json.dumps(
+                {
+                    "plot_directory": self.get_configuration()["directory"],
+                    "auto_handle_daq": True,
+                }
+            )
+            self.logger.info(
+                "Auto starting bandpass monitor with args: %s.", bandpass_args
+            )
+            self.start_bandpass_monitor(bandpass_args)
+            _wait_for_status(status="Bandpass Monitor", value="True")
+
+    def _is_integrated_channel_consumer_running(
+        self: DaqComponentManager, status: dict[str, Any] | None = None
+    ) -> bool:
+        """
+        Check if the INTEGRATED_CHANNEL_DATA consumer is running.
+
+        :param status: An optional status dictionary to check.
+
+        :return: True if the consumer is running, False otherwise.
+        """
+        if status is not None:
+            return bool(
+                str(["INTEGRATED_CHANNEL_DATA", 5])
+                in str(status.get("Running Consumers"))
+            )
+        return bool(
+            str(["INTEGRATED_CHANNEL_DATA", 5])
+            in str(json.loads(self.daq_status()).get("Running Consumers"))
+        )
+
+    def _is_bandpass_monitor_running(
+        self: DaqComponentManager, status: dict[str, Any] | None = None
+    ) -> bool:
+        """
+        Check if the bandpass monitor is running.
+
+        :param status: An optional status dictionary to check.
+
+        :return: True if the bandpass monitor is running, False otherwise.
+        """
+        if status is not None:
+            return bool(status.get("Bandpass Monitor"))
+        return bool(json.loads(self.daq_status()).get("Bandpass Monitor"))
+
+    def _reestablish_communication(
+        self: DaqComponentManager, configuration: str
+    ) -> None:
+        try:
+            response = self._daq_client.initialise(configuration)
+            self.logger.info(response["message"])
+            self._update_communication_state(CommunicationStatus.ESTABLISHED)
+            if not self._dedicated_bandpass_daq:
+                # This causes bandpass monitoring to start/stop/start
+                # if used in conjunction with the dedicated bandpass daq
+                # flag as comms becoming established/disabled
+                # enables/disables bandpass monitoring.
+                self.restart_daq_if_active()
+        except Exception as e:  # pylint: disable=broad-except
+            self.logger.error(
+                "Caught exception in start_communicating: %s. " + "Retrying in %s secs",
+                e,
+                self._daq_initialisation_retry_frequency,
+            )
 
     def stop_communicating(self: DaqComponentManager) -> None:
         """Break off communication with the DaqReceiver components."""
@@ -346,8 +511,6 @@ class DaqComponentManager(TaskExecutorComponentManager):
 
         :return: none
         """
-        if task_callback:
-            task_callback(status=TaskStatus.IN_PROGRESS)
         # Check data directory is in correct format, if not then reconfigure.
         # This delays the start call by a lot if SKUID isn't there.
         if not self._data_directory_format_adr55_compliant():
@@ -419,11 +582,8 @@ class DaqComponentManager(TaskExecutorComponentManager):
         self.logger.debug("Entering stop_daq")
         if task_callback:
             task_callback(status=TaskStatus.IN_PROGRESS)
-
         result_code, message = self._daq_client.stop_daq()
-
         self._started_event.clear()
-
         if task_callback:
             if result_code == ResultCode.OK:
                 task_callback(status=TaskStatus.COMPLETED)
@@ -441,14 +601,12 @@ class DaqComponentManager(TaskExecutorComponentManager):
         :param task_callback: Update task state, defaults to None
         :return: a task status and response message
         """
-        self.logger.debug("Entering daq_status")
         if task_callback:
             task_callback(status=TaskStatus.IN_PROGRESS)
 
         status = self._daq_client.get_status()
         if task_callback:
             task_callback(status=TaskStatus.COMPLETED)
-        self.logger.debug(f"Exiting daq_status with: {status}")
         return status
 
     @check_communicating
@@ -566,8 +724,6 @@ class DaqComponentManager(TaskExecutorComponentManager):
         config = self.get_configuration()
         nof_channels = int(config["nof_channels"])
 
-        if task_callback:
-            task_callback(status=TaskStatus.QUEUED)
         try:
             for response in self._daq_client.start_bandpass_monitor(argin):
                 x_bandpass_plot: np.ndarray | None = None
@@ -578,7 +734,6 @@ class DaqComponentManager(TaskExecutorComponentManager):
                 )
                 if task_callback is not None:
                     task_callback(
-                        status=TaskStatus(response["result_code"]),
                         result=response["message"],
                     )