+++ CHANGELOG.md
@@ -1,10 +1,14 @@
 # Version History
 
-## Unreleased
+## 1.0.0
+
+* [THORN-17] 1.0.0 release - all MCCS repos
+
+## 0.25.0
 
 * [SKB-761] Hardcode nof_antenna to number expected from library build.
 * [THORN-24] Update devices to use MccsBaseDevice for mode inheritance.
-* [SKB-765] Add ConfigureStationForCalibration command, which handles DAQ configuration for calibration. This functionality was pulled from AcquireDataForCalibration.
+* [THORN-27] Add `BandpassDaq` property to DaqReceiver. Where this is `True` these Daqs will automatically attempt to keep bandpass monitoring running.
 
 ## 0.24.0